[Refactor] 全体リファクタリング・コードレビュー修正
- pty.rs: io::Error::new を io::Error::other に統一、u16→i16 キャストを安全なクランプに変更
- build.rs: clippy 推奨の contains/div_ceil/repeat_n/>= に書き換え
- render.rs: draw_text に clippy::too_many_arguments サプレス追加
- tsf.rs: 不要な as u32 キャスト除去、arc_with_non_send_sync サプレス追加
- term.rs: 余分な空行を除去

+++ src/tsf.rs
@@ -351,7 +351,7 @@ impl ITextStoreACP_Impl for TextStore_Impl {
         // composition 中は preedit 末尾が挿入点
         let insert_acp = base_acp.saturating_add(preedit_len);
 
-        if dwflags & TF_IAS_QUERYONLY.0 as u32 != 0 {
+        if dwflags & TF_IAS_QUERYONLY.0 != 0 {
             unsafe {
                 if !pacpstart.is_null() { *pacpstart = insert_acp; }
                 if !pacpend.is_null() { *pacpend = insert_acp; }
@@ -583,6 +583,7 @@ pub fn setup_tsf(
     let client_id = unsafe { thread_mgr.Activate()? };
     let doc_mgr = unsafe { thread_mgr.CreateDocumentMgr()? };
 
+    #[allow(clippy::arc_with_non_send_sync)] // COM オブジェクトは Send/Sync ではないが Arc<Mutex> で保護
     let shared_sink = Arc::new(Mutex::new(SharedSink { sink: None, mask: 0 }));
     let composition = Arc::new(Mutex::new(CompositionState::new()));
     let text_store = TextStore::new(
